interchange: add new protobuf decode module
